Smart-Parking-App/backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0

channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1

gunicorn==25.3.0
whitenoise==6.12.0

psycopg2-binary==2.9.11
redis==7.1.0

python-dotenv==1.1.1
PyMySQL==1.1.2